Create Caesar's Cypher Challenge
I know I am a little late to the Capture the flag challenge but I figure learning how to do this will help when the next one comes. This is not an official Submissioon for PicoCTF2019 - it's just me attempting to prove that I worked my way through the exercise.

This BTW is my Edit to the Caesar's Cypher that is in the book.

https://primer.picoctf.com/#_cryptography

Very easy. all I had to do was edit the modulo to redefine the upper boundary.